rename project to ng-config, as previous name was the name of the Angular Events conference :D
