+++ projects/app/src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component, Inject } from "@angular/core";
 import { AUTH_SETTINGS } from "projects/ng-auth/src/lib/tokens";
-import { APP_SETTINGS, IAUTH_SETTINGS } from "projects/ng-conf/src/lib/tokens";
-import { IAppSettings } from 'projects/ng-conf/src/lib/types';
+import { APP_SETTINGS, IAUTH_SETTINGS } from "projects/ng-config/src/lib/tokens";
+import { IAppSettings } from 'projects/ng-config/src/lib/types';
 
 
 @Component({
