+++ projects/ng-config/README.md
@@ -1,0 +1,6 @@
+# NgConfig
+
+This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 17.3.0.
+
+It's a library that provides runtime configuration objects after retrieving from a remote config file.
+
