+++ projects/ng-config/ng-package.json
