+++ projects/ng-config/src/lib/settings.service.ts
