+++ projects/ng-config/src/lib/types.ts
