+++ projects/ng-config/src/public-api.ts
