+++ projects/ng-config/tsconfig.lib.json
