+++ projects/ng-config/tsconfig.lib.prod.json
